Homeworkforbranches

+++ app/Http/Controllers/BranchController.php
@@ -22,11 +22,20 @@ class BranchController extends ApiController
     }
 
 
-    public function show(Branch $branch): JsonResponse
+    public function all(): JsonResponse
     {
-        return $this->successResponse($this->service->show($branch));
+        return $this->successResponse($this->service->getAll());
     }
 
+
+    public function show(Branch $branch): JsonResponse
+    {
+        dd(auth('api')->user()->branches);
+        if(!in_array($branch->id, auth('api')->user()->branches->pluck('id')->toArray(), true)){
+            return  $this->errorResponse('Forbidden',403);
+        }
+        return $this->successResponse($this->service->show($branch));
+    }
 
     /**
      * @throws ValidatorException
@@ -45,11 +54,13 @@ class BranchController extends ApiController
      */
     public function update(BranchRequest $request, Branch $branch)
     {
+        if(!in_array($branch->id, auth('api')->user()->branches->pluck('id')->toArray(), true)){
+            return  $this->errorResponse('Forbidden',403);
+        }
 
         if($request->validated()){
             return $this->successResponse( $this->service->update($branch, $request->validated()));
         }
-
 
         return $this->errorResponse('Validation Failed', Response::HTTP_UNPROCESSABLE_ENTITY);
     }
@@ -57,6 +68,10 @@ class BranchController extends ApiController
 
     public function destroy(Branch $branch): JsonResponse
     {
+        if(!in_array($branch->id, auth('api')->user()->branches->pluck('id')->toArray(), true)){
+            return  $this->errorResponse('Forbidden',403);
+        }
+
         return $this->successResponse((array)$this->service->delete($branch),Response::HTTP_NO_CONTENT);
     }
 }
